Stream backed writers passed to XMLWriter must use UTF-8 encoding
The XMLWriter writes a character stream and outputs 

<?xml version='1.0'?>

as XML declaration. The encoding declaration is missing, so encoding
must comply with UTF-8.

This change modifies the places where the output is written into a
file to use UTF-8 encoding. Without this change the default charset
of the platform is used, resulting in illegal bytesequences in the
output file.