WIP: take into account global gitignore
Take into account exclusions from the global gitignore file if it is
present in the default location. It's kind of icky to read environment
variables in code called from an enum constructor, but I didn't see a
more elegant solution yet either. Relatedly, I have not yet created a
unit test (or made sure that a global gitignore will not interfere with
a global gitignore being present), hence the WIP status.

https://git-scm.com/docs/gitignore